시간 크기 Uint64 -> Uint32로 변경, 헤더 내용 조정 ( 추후 분할 패킷 작업 ) (#16)
* feat: 패킷 헤더 플래그 변경

* fix: 부호 비트 안맞는 문제 해결

* feat: 헤더 패딩 4로 고정

+++ DUBUFrameWork/PendingPacket.h
@@ -9,8 +9,8 @@ namespace DUBU
 	struct PendingPacket
 	{
 		struct OverlappedPacketBuffer* buffer = nullptr;
-		Uint64   timeStamp = 0;
-		Uint32  sequenceNo = 0;
-		bool    isSent = false;
+		Uint32 timeStamp = 0;
+		Uint32 sequenceNo = 0;
+		bool isSent = false;
 	};
 };
